+++ tests/test_yaralyze.py
@@ -15,7 +15,7 @@ import pytest
 from yaralyzer.output.console import console
 from yaralyzer.util.constants import NO_TIMESTAMPS_OPTION, YARALYZE
 from yaralyzer.util.helpers.file_helper import files_in_dir, load_file
-from yaralyzer.util.helpers.shell_helper import compare_export_to_file
+from yaralyzer.util.helpers.shell_helper import compare_export_to_file, safe_args
 from yaralyzer.util.helpers.string_helper import line_count
 from yaralyzer.util.logging import log, log_bigly
 
@@ -123,4 +123,4 @@ def _compare_to_fixture(file_to_scan: str | Path, *args):
 
 
 def _build_shell_cmd(file_path: str | Path, *args) -> list[str]:
-    return [YARALYZE, str(file_path), *[str(arg) for arg in args]]
+    return [YARALYZE] + safe_args([file_path, *args])